Fix the URL quoting in escapeImages()
Previously, URLs that had a colon in the scheme part, such as
'http://', would be mangled (e.g. to 'http%3A//') and the
images referred to by such URLs would not be displayed.  This
new urlQuote function will selectively escape characters in
the the path and query parts instead.

The function is based on an answer to this stackoverflow
question:

  http://stackoverflow.com/questions/120951/how-can-i-normalize-a-url-in-python